tested code and fixed getPlayList function
