--- alvin-notes.md
@@ -1,1 +1,0 @@
-my secrets
