+++ server/server.js
@@ -1,22 +1,14 @@
-const express = require('express');
 require('dotenv').config();
+const express = require('express');
 const app = express();
 const path = require('path');
-const uri =process.env.MONGO_URI;
+const uri =`mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASS}@cluster.0own7hn.mongodb.net/
+`;
+const db = require('./model.js');
 const mongoose = require('mongoose');
 const cors = require('cors');
-
-// import controllers
-const UserController = require('./controllers/userController.js');
-const GroupController = require('./controllers/groupController.js');
-const PlaylistController = require('./controllers/playlistController.js');
-
-// import routes
-const userRouter = require('./routes/user.js');
-const groupRoute = require('./routes/group.js');
-
-
-// const PORT = 3000;
+const mainController = require('./controller.js');
+const PORT = 3000;
 
 mongoose.connect(uri, {
   useNewUrlParser: true,
@@ -36,32 +28,37 @@ app.get('/', (req, res) => {
   return res.status(200).sendFile(path.join(__dirname, '../client/index.html'));
 });
 
-app.use('/user', userRouter)
-
 // createUser
 app.post('/user', mainController.createUser, (req, res) => {
   return res.status(200).json(res.locals.newUser);
 });
-app.post('/group', mainController.createGroup, (req, res) => {
-  return res.status(200).json(res.locals.newGroup);
-});
-app.post('/playlist', mainController.createPlaylist, (req, res) => {
-  return res.status(200).json(res.locals.newPlaylist);
-});
-app.post('/addGroup', mainController.addGroupToUser, (req, res) => {
-  return res.status(200).json(res.locals.groupArr);
-});
-// in body add user object that you want to get
 app.get('/user/:username', mainController.getUser, (req, res) => {
   return res.status(200).json(res.locals.foundUser);
 });
-//localhost:3000/group/1
+
+//group functions
+app.post('/group', mainController.createGroup, (req, res) => {
+  return res.status(200).json(res.locals.newGroup);
+});
 app.get('/group/:groupID', mainController.getGroup, (req, res) => {
   return res.status(200).json(res.locals.foundGroup);
 });
-app.get('/playlist', mainController.getPlaylist, (req, res) => {
+
+app.post('/playlist', mainController.createPlaylist, (req, res) => {
+  return res.status(200).json(res.locals.newPlaylist);
+});
+app.get('/playlist/:groupID', mainController.getPlaylist, (req, res) => {
   return res.status(200).json(res.locals.foundPlaylist);
 });
+
+app.patch('/user/:groups', mainController.addGroupToUser, (req, res) => {
+  return res.status(200).json(res.locals.groupArr);
+});
+// in body add user object that you want to get
+
+//localhost:3000/group/1
+
+
 // // DELETE ALL GROUPS,PLAYLISTS, USERS
   // deleteAllGroups
 app.delete('/group', mainController.deleteAllGroups, (req,res)=>{
@@ -106,10 +103,10 @@ app.use((err, req, res, next) => {
   return res.status(errorObj.status).json(errorObj.message);
 });
 
-// app.listen(PORT, () => console.log(`Listening on PORT: ${PORT}`));
+app.listen(PORT, () => console.log(`Listening on PORT: ${PORT}`));
 
-app.listen(3000, () => {
-  console.log('server started on 3000', process.env.NODE_ENV);
-});
+// app.listen(PORT, () => {
+//   console.log('server started on 3000');
+// });
 
 module.exports = app;
